Add support for BME280 sensor
From the BME280 datasheet(Section 5.2), BME280 is register compatible
with BMP280, and the pressure & temperature control and readout is
identical, with a few exceptions that are considered below.

1. BME280 has a ChipID of 0x60 (Code has been updated to accept this ChipID)

2. t_sb has a different map. The current code uses a value of t_sb of 
StandbyTime63ms (0x01) which is the same across both sensors

3. Pressure resolution might differ on filter settings for BME280

-- BMP280 resolution only depends on osrs_p ,which is set to Oversamp16x
(0x05) corresponding to an oversampling of 16x. This sets the pressure
resolution to 20 bits.

-- BME280 resolution depends if the filter is on or off. In the current
code, filter is set to FilterCoeff16(0x04), which corresponds to a
filter coefficient of 16. This sets the pressure resolution to 20 bits.

Pressure resolution is set to 20 bits on both sensors

4. Temperature resolution

Temperature resolution is set to 20 bits on both sensors following a
similar reasoning as point 3 above.


Tested on BME280 module(GY-BM) wired directly to raspberry pi with the
following pins and using  I2C address of 0x76.
- Pin 1 : 3.3v
- Pin 3 : SDA
- Pin 5 : SCL
- Pin 9 : GND

NO IMU module was connected.

Signed-off-by: Amit Dhingra <mechanicalamit@gmail.com>